chromium/third_party/blink/web_tests/web-components-v0-only/fast/html-imports/xhr.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<link id="client" rel="import" href="import-client.html">
</head>
<body>
<script>
description("Ensure XHR document doesn't load imports");
window.jsTestIsAsync = true;

var xhr = new XMLHttpRequest();
xhr.open("GET", "resources/root.html", true);
xhr.responseType = "document";
xhr.onload = function(e) {
   link = xhr.response.head.querySelector("#hello");
   shouldBeNull("link.import");
   finishJSTest();
}

xhr.send();

</script>
</body>
</html>
